Rename non-core effect files
