+++ server/game/core/ongoingEffect/effectImpl/Restriction.ts
@@ -1,7 +1,7 @@
 import { AbilityContext } from '../../ability/AbilityContext';
 import Player from '../../Player';
 import { OngoingEffectValueWrapper } from './OngoingEffectValueWrapper';
-import { restrictionDsl } from '../../../effects/RestrictionDsl';
+import { restrictionDsl } from '../../../ongoingEffects/RestrictionDsl';
 import type Card from '../../card/Card';
 
 const leavePlayTypes = new Set(['discardFromPlay', 'returnToHand', 'returnToDeck', 'removeFromGame']);
